+++ src/realtimecontroller/DefaultRealtimeController.ts
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import DataMessage from '../datamessage/DataMessage';
-import DefaultDeviceController from '../devicecontroller/DefaultDeviceController';
 import MediaStreamBroker from '../mediastreambroker/MediaStreamBroker';
 import DefaultTranscriptionController from '../transcript/DefaultTranscriptionController';
 import TranscriptionController from '../transcript/TranscriptionController';
@@ -468,10 +467,12 @@ export default class DefaultRealtimeController implements RealtimeController {
     if (attendeeIdRemote !== attendeeIdLocal) {
       return mutedRemote;
     }
-    // This is a workaround for Chime App dial in feature
-    // @ts-ignore
-    if (!((this.mediaStreamBroker as DefaultDeviceController).activeDevices['audio'])) {
-      return mutedRemote
+    // This is a workaround to check if no audio input then just use the remote value
+    if (
+      // @ts-ignore
+      !('activeDevices' in this.mediaStreamBroker && this.mediaStreamBroker.activeDevices['audio'])
+    ) {
+      return mutedRemote;
     }
     return mutedLocal;
   }